Upated float to double

+++ proto/c/include/transcoder.h
@@ -64,8 +64,9 @@ size_t EncodePowerMeasurement(uint32_t ts, uint32_t logger_id,
  * @return Number of bytes in @p buffer
 */
 size_t EncodeTeros12Measurement(uint32_t ts, uint32_t logger_id,
-                                uint32_t cell_id, float vwc_raw, float vwc_adj,
-                                float temp, uint32_t ec, uint8_t *buffer);
+                                uint32_t cell_id, double vwc_raw,
+                                double vwc_adj, double temp, uint32_t ec,
+                                uint8_t *buffer);
 
 /**
  * @brief Decodes a response message
